Refactor Azure DevOps run ID coordination
Split the coordinator into focused partial-class files for leases, participants, run ID files, file I/O, and coordination models.
