+++ src/Platform/Microsoft.Testing.Extensions.AzureDevOpsReport/AzureDevOpsRunIdCoordinator.cs
@@ -1,8 +1,5 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System.Text.Json;
-using System.Text.Json.Serialization;
 
 using Microsoft.Testing.Extensions.AzureDevOpsReport.Resources;
 using Microsoft.Testing.Platform.Helpers;
@@ -10,20 +7,9 @@ using Microsoft.Testing.Platform.Logging;
 
 namespace Microsoft.Testing.Extensions.AzureDevOpsReport;
 
-internal sealed class AzureDevOpsRunIdCoordinator
+internal sealed partial class AzureDevOpsRunIdCoordinator
 {
     private const string CoordinationFilePrefix = "azdo-runid";
-
-    /// <summary>
-    /// Build id recorded for a legacy plain-PID lease, which carries no build context of its own.
-    /// </summary>
-    private const int LegacyLeaseBuildId = 0;
-    private static readonly JsonSerializerOptions JsonSerializerOptions = new()
-    {
-        DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull,
-    };
-
-    private static readonly UTF8Encoding Utf8EncodingWithoutBom = new(encoderShouldEmitUTF8Identifier: false);
 
     private readonly IFileSystem _fileSystem;
     private readonly ITask _task;
@@ -209,431 +195,4 @@ internal sealed class AzureDevOpsRunIdCoordinator
             TryDeleteFile(coordinatedRun.RunIdFilePath);
         }
     }
-
-    /// <summary>
-    /// Returns whether at least one of the given participants belongs to a process that is still running.
-    /// </summary>
-    /// <remarks>
-    /// Distinguishes "a peer is still publishing" from "a lease is lying around that nobody can vouch for".
-    /// Only the former justifies holding the run open past the short grace period.
-    /// </remarks>
-    private bool AnyParticipantProcessAlive(string[] participantFiles)
-    {
-        foreach (string participantFile in participantFiles)
-        {
-            AzureDevOpsLeaseFile? lease = ReadLease(participantFile).Lease;
-            int? processId = lease?.ProcessId ?? TryGetPid(participantFile);
-
-            // Our own process is always alive, so counting it would mean waiting for ourselves forever.
-            if (processId is { } participantProcessId
-                && participantProcessId != _environment.ProcessId
-                && IsProcessAlive(participantProcessId))
-            {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    private string[] CleanupStaleParticipants(string[] participantFiles)
-    {
-        var activeParticipants = new List<string>();
-
-        foreach (string participantFile in participantFiles)
-        {
-            LeaseReadResult result = ReadLease(participantFile);
-
-            // Treat a participant whose file we couldn't read (e.g. mid-write) as active to avoid
-            // racing with a process that is still updating its lease.
-            if (result.Status is LeaseFileStatus.TransientReadError or LeaseFileStatus.Active)
-            {
-                AzureDevOpsLeaseFile? lease = result.Lease;
-                if (lease is null || (lease.ExpiresAt > _clock.UtcNow && IsProcessAlive(lease.ProcessId)))
-                {
-                    activeParticipants.Add(participantFile);
-                    continue;
-                }
-            }
-            else if (result.Status == LeaseFileStatus.Expired
-                && result.Lease is { } expiredLease
-                && IsProcessAlive(expiredLease.ProcessId))
-            {
-                // Lease has expired according to wall-clock but the participant process is still
-                // running (its renewal may simply be stuck). Keep waiting rather than deleting.
-                activeParticipants.Add(participantFile);
-                continue;
-            }
-            else if (result.Status == LeaseFileStatus.NotFound)
-            {
-                continue;
-            }
-            else if (TryGetPid(participantFile) is int processId && IsProcessAlive(processId))
-            {
-                activeParticipants.Add(participantFile);
-                continue;
-            }
-
-            TryDeleteFile(participantFile);
-        }
-
-        return [.. activeParticipants];
-    }
-
-    private static bool IsProcessAlive(int processId)
-    {
-        try
-        {
-            using var process = Process.GetProcessById(processId);
-            return !process.HasExited;
-        }
-        catch (ArgumentException)
-        {
-            return false;
-        }
-        catch (InvalidOperationException)
-        {
-            return false;
-        }
-    }
-
-    [UnconditionalSuppressMessage("ReflectionAnalysis", "IL2026", Justification = "The coordination payload type is internal, fixed, and controlled by this extension.")]
-    [UnconditionalSuppressMessage("Aot", "IL3050", Justification = "The coordination payload type is internal, fixed, and controlled by this extension.")]
-    private async Task<AzureDevOpsRunIdFile?> WaitForRunIdFileAsync(string runIdFilePath, string ownerFilePath, int buildId, CancellationToken cancellationToken)
-    {
-        DateTimeOffset joinerDeadline = _clock.UtcNow + _options.CoordinationJoinerMaxWaitTime;
-
-        for (int attempt = 0; ; attempt++)
-        {
-            if (_fileSystem.ExistFile(runIdFilePath))
-            {
-                try
-                {
-                    string content = await _fileSystem.ReadAllTextAsync(runIdFilePath).ConfigureAwait(false);
-                    AzureDevOpsRunIdFile? runIdFile = JsonSerializer.Deserialize<AzureDevOpsRunIdFile>(content, JsonSerializerOptions);
-                    if (runIdFile is not null)
-                    {
-                        // A run-id file whose expiry has passed is only stale if nobody owns it any more.
-                        // An owner that does not renew its lease (a run orchestrator holds one for the whole
-                        // orchestration) is still publishing into that run, so deleting the file here would
-                        // send every joiner off to create a second run for the same build.
-                        if (runIdFile.BuildId != buildId
-                            || (runIdFile.ExpiresAt <= _clock.UtcNow && !IsOwnerStillAlive(ReadLease(ownerFilePath))))
-                        {
-                            TryDeleteFile(runIdFilePath);
-                        }
-                        else if (runIdFile.BuildId == buildId)
-                        {
-                            return runIdFile;
-                        }
-                    }
-                }
-                catch (IOException)
-                {
-                }
-                catch (UnauthorizedAccessException)
-                {
-                }
-                catch (JsonException)
-                {
-                }
-            }
-
-            // After the base retry budget is exhausted, only keep waiting as long as the owner lease
-            // still looks active (or temporarily unreadable). A long CreateTestRunAsync can outlast
-            // CoordinationReadRetryCount * CoordinationReadRetryDelay, but we still want to bound the
-            // wait by CoordinationJoinerMaxWaitTime so a crashed owner doesn't stall joiners forever.
-            if (attempt >= _options.CoordinationReadRetryCount)
-            {
-                if (_clock.UtcNow >= joinerDeadline)
-                {
-                    return null;
-                }
-
-                LeaseReadResult ownerLease = ReadLease(ownerFilePath);
-                bool ownerStillActive = ownerLease.Status is LeaseFileStatus.Active or LeaseFileStatus.TransientReadError
-                    || (ownerLease.Status == LeaseFileStatus.Expired && IsOwnerStillAlive(ownerLease));
-
-                if (!ownerStillActive)
-                {
-                    return null;
-                }
-            }
-
-            await _task.Delay(_options.CoordinationReadRetryDelay, cancellationToken).ConfigureAwait(false);
-        }
-    }
-
-    [UnconditionalSuppressMessage("ReflectionAnalysis", "IL2026", Justification = "The coordination payload type is internal, fixed, and controlled by this extension.")]
-    [UnconditionalSuppressMessage("Aot", "IL3050", Justification = "The coordination payload type is internal, fixed, and controlled by this extension.")]
-    private async Task WriteParticipantLeaseAsync(string participantFilePath, int buildId, CancellationToken cancellationToken)
-    {
-        AzureDevOpsLeaseFile lease = CreateLeaseFile(buildId);
-        if (await TryWriteLeaseFileAsync(participantFilePath, lease, overwrite: false, cancellationToken).ConfigureAwait(false))
-        {
-            return;
-        }
-
-        await WriteLeaseFileAsync(participantFilePath, lease, overwrite: true, cancellationToken).ConfigureAwait(false);
-    }
-
-    [UnconditionalSuppressMessage("ReflectionAnalysis", "IL2026", Justification = "The coordination payload type is internal, fixed, and controlled by this extension.")]
-    [UnconditionalSuppressMessage("Aot", "IL3050", Justification = "The coordination payload type is internal, fixed, and controlled by this extension.")]
-    private async Task<bool> TryAcquireOwnerAsync(string ownerFilePath, int buildId, CancellationToken cancellationToken)
-    {
-        AzureDevOpsLeaseFile lease = CreateLeaseFile(buildId);
-        if (await TryWriteLeaseFileAsync(ownerFilePath, lease, overwrite: false, cancellationToken).ConfigureAwait(false))
-        {
-            return true;
-        }
-
-        LeaseReadResult existing = ReadLease(ownerFilePath);
-
-        // If the file exists but we couldn't read it (likely partial write from the current owner),
-        // refuse to take over — otherwise we'd race and create a duplicate Azure DevOps run.
-        if (existing.Status == LeaseFileStatus.TransientReadError)
-        {
-            return false;
-        }
-
-        if (existing.Status is LeaseFileStatus.NotFound
-            || (existing.Status == LeaseFileStatus.Expired && !IsOwnerStillAlive(existing)))
-        {
-            TryDeleteFile(ownerFilePath);
-            return await TryWriteLeaseFileAsync(ownerFilePath, lease, overwrite: false, cancellationToken).ConfigureAwait(false);
-        }
-
-        return false;
-    }
-
-    /// <summary>
-    /// Returns whether the process that wrote a lease is still running.
-    /// </summary>
-    /// <remarks>
-    /// Liveness, not wall-clock expiry, is the authoritative signal that an owner is still in charge. Not
-    /// every owner renews its lease: a run orchestrator holds one for the whole orchestration without a
-    /// natural point to refresh it, so on a build longer than
-    /// <see cref="AzureDevOpsTestResultsPublisherOptions.CoordinationFileExpiration"/> its lease goes stale
-    /// while it is very much still publishing. Taking over then would create a second Azure DevOps run for
-    /// the same build. Refusing to take over from a live process errs towards one run rather than two,
-    /// which is also how <see cref="CleanupStaleParticipants"/> already treats participants.
-    /// <para>
-    /// The legacy plain-PID lease format is excluded: <see cref="ReadLease"/> deliberately reports it as
-    /// expired so it can be replaced, and it carries no build id to tell us which run it belongs to. A
-    /// reused pid would otherwise let a stale file from an older version block takeover indefinitely.
-    /// </para>
-    /// </remarks>
-    private static bool IsOwnerStillAlive(LeaseReadResult lease)
-        => lease.Lease is { BuildId: not LegacyLeaseBuildId } ownerLease && IsProcessAlive(ownerLease.ProcessId);
-
-    private AzureDevOpsLeaseFile CreateLeaseFile(int buildId)
-        => new(_environment.ProcessId, buildId, _clock.UtcNow + _options.CoordinationFileExpiration);
-
-    [UnconditionalSuppressMessage("ReflectionAnalysis", "IL2026", Justification = "The coordination payload type is internal, fixed, and controlled by this extension.")]
-    [UnconditionalSuppressMessage("Aot", "IL3050", Justification = "The coordination payload type is internal, fixed, and controlled by this extension.")]
-    private LeaseReadResult ReadLease(string path)
-    {
-        if (!_fileSystem.ExistFile(path))
-        {
-            return new LeaseReadResult(LeaseFileStatus.NotFound, null);
-        }
-
-        try
-        {
-            string content = _fileSystem.ReadAllText(path);
-            AzureDevOpsLeaseFile? lease = JsonSerializer.Deserialize<AzureDevOpsLeaseFile>(content, JsonSerializerOptions);
-            if (lease is not null)
-            {
-                return new LeaseReadResult(
-                    lease.ExpiresAt > _clock.UtcNow ? LeaseFileStatus.Active : LeaseFileStatus.Expired,
-                    lease);
-            }
-
-            if (int.TryParse(content, NumberStyles.Integer, CultureInfo.InvariantCulture, out int processId))
-            {
-                // Legacy plain-PID lease format: treat as expired so the caller can take over. The sentinel
-                // build id marks it as carrying no build context, which is what keeps IsOwnerStillAlive from
-                // treating a reused pid as a live owner.
-                return new LeaseReadResult(LeaseFileStatus.Expired, new AzureDevOpsLeaseFile(processId, LegacyLeaseBuildId, DateTimeOffset.MinValue));
-            }
-
-            // The file exists but neither parser yielded a usable value. It might be mid-write —
-            // surface this as a transient read error so the caller doesn't race the writer.
-            return new LeaseReadResult(LeaseFileStatus.TransientReadError, null);
-        }
-        catch (IOException)
-        {
-        }
-        catch (UnauthorizedAccessException)
-        {
-        }
-        catch (JsonException)
-        {
-        }
-
-        return new LeaseReadResult(LeaseFileStatus.TransientReadError, null);
-    }
-
-    [UnconditionalSuppressMessage("ReflectionAnalysis", "IL2026", Justification = "The coordination payload type is internal, fixed, and controlled by this extension.")]
-    [UnconditionalSuppressMessage("Aot", "IL3050", Justification = "The coordination payload type is internal, fixed, and controlled by this extension.")]
-    private async Task WriteRunIdFileAsync(string runIdFilePath, AzureDevOpsPublishConfiguration configuration, int runId, CancellationToken cancellationToken)
-        => await WriteJsonFileAsync(runIdFilePath, new AzureDevOpsRunIdFile(runId, configuration.BuildId, configuration.CollectionUri, configuration.Project, _clock.UtcNow + _options.CoordinationFileExpiration), overwrite: true, cancellationToken).ConfigureAwait(false);
-
-    /// <summary>
-    /// Publishes the run id for peers to join, tolerating a failure to write the file.
-    /// </summary>
-    /// <remarks>
-    /// The run already exists in Azure DevOps at this point, and this process is the only one that can
-    /// close it. Letting anything propagate here would unwind into the caller's cleanup, which deletes the
-    /// leases and never hands back the run, stranding it in "InProgress" forever. So nothing escapes:
-    /// not a file-system failure, not cancellation, and not a logger provider that throws while reporting
-    /// it. Losing the file costs peers the ability to join this run, which is degraded but recoverable;
-    /// losing the run itself is not.
-    /// </remarks>
-    private async Task TryWriteRunIdFileAsync(string runIdFilePath, AzureDevOpsPublishConfiguration configuration, int runId, CancellationToken cancellationToken)
-    {
-        try
-        {
-            await WriteRunIdFileAsync(runIdFilePath, configuration, runId, cancellationToken).ConfigureAwait(false);
-        }
-        catch (Exception ex)
-        {
-            TryLogWarning($"{AzureDevOpsResources.AzureDevOpsLivePublishingFailedToWriteCoordinationFile} {runIdFilePath}: {ex.Message}");
-        }
-    }
-
-    /// <summary>
-    /// Logs a warning, swallowing any failure from the logging providers.
-    /// </summary>
-    /// <remarks>
-    /// The aggregate logger invokes each provider directly, so a failing provider propagates. Every caller
-    /// here is already recovering from something, and letting a diagnostic replace the failure it was
-    /// describing would lose both.
-    /// </remarks>
-    private void TryLogWarning(string message)
-    {
-        try
-        {
-            _logger.LogWarning(message);
-        }
-        catch (Exception)
-        {
-            // There is nowhere left to report this: the diagnostic logger is the fallback sink.
-        }
-    }
-
-    [UnconditionalSuppressMessage("ReflectionAnalysis", "IL2026", Justification = "The coordination payload type is internal, fixed, and controlled by this extension.")]
-    [UnconditionalSuppressMessage("Aot", "IL3050", Justification = "The coordination payload type is internal, fixed, and controlled by this extension.")]
-    private Task WriteLeaseFileAsync(string path, AzureDevOpsLeaseFile payload, bool overwrite, CancellationToken cancellationToken)
-        => WriteJsonFileAsync(path, payload, overwrite, cancellationToken);
-
-    [UnconditionalSuppressMessage("ReflectionAnalysis", "IL2026", Justification = "The coordination payload type is internal, fixed, and controlled by this extension.")]
-    [UnconditionalSuppressMessage("Aot", "IL3050", Justification = "The coordination payload type is internal, fixed, and controlled by this extension.")]
-    private async Task<bool> TryWriteLeaseFileAsync(string path, AzureDevOpsLeaseFile payload, bool overwrite, CancellationToken cancellationToken)
-    {
-        try
-        {
-            await WriteLeaseFileAsync(path, payload, overwrite, cancellationToken).ConfigureAwait(false);
-            return true;
-        }
-        catch (IOException)
-        {
-            return false;
-        }
-        catch (UnauthorizedAccessException)
-        {
-            return false;
-        }
-    }
-
-    [UnconditionalSuppressMessage("ReflectionAnalysis", "IL2026", Justification = "The coordination payload type is internal, fixed, and controlled by this extension.")]
-    [UnconditionalSuppressMessage("Aot", "IL3050", Justification = "The coordination payload type is internal, fixed, and controlled by this extension.")]
-    private async Task WriteJsonFileAsync<TPayload>(string path, TPayload payload, bool overwrite, CancellationToken cancellationToken)
-    {
-        string json = JsonSerializer.Serialize(payload, JsonSerializerOptions);
-        using IFileStream stream = _fileSystem.NewFileStream(path, overwrite ? FileMode.Create : FileMode.CreateNew, FileAccess.Write, FileShare.Read);
-        using StreamWriter writer = new(stream.Stream, Utf8EncodingWithoutBom, 1024, leaveOpen: true);
-#if NET
-        await writer.WriteAsync(json.AsMemory(), cancellationToken).ConfigureAwait(false);
-        await writer.FlushAsync(cancellationToken).ConfigureAwait(false);
-#else
-        await writer.WriteAsync(json).ConfigureAwait(false);
-        await writer.FlushAsync().ConfigureAwait(false);
-#endif
-    }
-
-    private void TryDeleteFile(string path)
-    {
-        try
-        {
-            if (_fileSystem.ExistFile(path))
-            {
-                _fileSystem.DeleteFile(path);
-            }
-        }
-        catch (IOException ex)
-        {
-            TryLogWarning($"{AzureDevOpsResources.AzureDevOpsLivePublishingFailedToDeleteCoordinationFile} {path}: {ex.Message}");
-        }
-        catch (UnauthorizedAccessException ex)
-        {
-            TryLogWarning($"{AzureDevOpsResources.AzureDevOpsLivePublishingFailedToDeleteCoordinationFile} {path}: {ex.Message}");
-        }
-    }
-
-    private static int? TryGetPid(string participantFile)
-    {
-        string fileName = Path.GetFileNameWithoutExtension(participantFile);
-        int lastSeparator = fileName.LastIndexOf('.');
-        return lastSeparator < 0
-            ? null
-            : int.TryParse(fileName[(lastSeparator + 1)..], NumberStyles.Integer, CultureInfo.InvariantCulture, out int processId)
-                ? processId
-                : null;
-    }
-
-    private static string GetOwnerFileName(int buildId)
-        => $"{CoordinationFilePrefix}.{buildId}.owner";
-
-    private static string GetParticipantFileName(int buildId, int processId)
-        => $"{CoordinationFilePrefix}.{buildId}.participant.{processId}.json";
-
-    private static string GetParticipantSearchPattern(int buildId)
-        => $"{CoordinationFilePrefix}.{buildId}.participant.*.json";
-
-    private static string GetRunIdFileName(int buildId)
-        => $"{CoordinationFilePrefix}.{buildId}.json";
 }
-
-internal sealed record AzureDevOpsCoordinatedRun(
-    int RunId,
-    bool IsOwner,
-    int BuildId,
-    string ResultsDirectory,
-    string RunIdFilePath,
-    string OwnerFilePath,
-    string ParticipantFilePath)
-{
-    /// <summary>
-    /// Gets a value indicating whether the run was created by an ancestor process, which also owns closing it.
-    /// </summary>
-    /// <remarks>
-    /// Declared as a property rather than a positional parameter so that adding it does not change the
-    /// record's constructor and deconstructor signatures. An inherited run has no coordination files of
-    /// its own: the ancestor decides when the run ends, so there is no owner to elect and no participant
-    /// set to drain.
-    /// </remarks>
-    public bool IsInherited { get; init; }
-}
-
-internal sealed record AzureDevOpsRunIdFile(
-    [property: JsonPropertyName("runId")] int RunId,
-    [property: JsonPropertyName("buildId")] int BuildId,
-    [property: JsonPropertyName("collectionUri")] string CollectionUri,
-    [property: JsonPropertyName("project")] string Project,
-    [property: JsonPropertyName("expiresAt")] DateTimeOffset ExpiresAt);
-
-internal sealed record AzureDevOpsLeaseFile(
-    [property: JsonPropertyName("processId")] int ProcessId,
-    [property: JsonPropertyName("buildId")] int BuildId,
-    [property: JsonPropertyName("expiresAt")] DateTimeOffset ExpiresAt);
